Move SearchResultsView into a separate file
